--- alembic/versions/5dafe949fa14_init_database.py
@@ -1,32 +1,0 @@
-"""Init database
-
-Revision ID: 5dafe949fa14
-Revises: c36d9883e62f
-Create Date: 2025-06-16 08:38:06.784004
-
-"""
-from typing import Sequence, Union
-
-from alembic import op
-import sqlalchemy as sa
-
-
-# revision identifiers, used by Alembic.
-revision: str = '5dafe949fa14'
-down_revision: Union[str, None] = 'c36d9883e62f'
-branch_labels: Union[str, Sequence[str], None] = None
-depends_on: Union[str, Sequence[str], None] = None
-
-
-def upgrade() -> None:
-    """Upgrade schema."""
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.drop_column('tags', 'description')
-    # ### end Alembic commands ###
-
-
-def downgrade() -> None:
-    """Downgrade schema."""
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.add_column('tags', sa.Column('description', sa.TEXT(), autoincrement=False, nullable=True))
-    # ### end Alembic commands ###
