--- alembic/versions/c36d9883e62f_init_database.py
@@ -1,32 +1,0 @@
-"""Init database
-
-Revision ID: c36d9883e62f
-Revises: 0ebebf4fd523
-Create Date: 2025-06-13 17:44:40.133894
-
-"""
-from typing import Sequence, Union
-
-from alembic import op
-import sqlalchemy as sa
-
-
-# revision identifiers, used by Alembic.
-revision: str = 'c36d9883e62f'
-down_revision: Union[str, None] = '0ebebf4fd523'
-branch_labels: Union[str, Sequence[str], None] = None
-depends_on: Union[str, Sequence[str], None] = None
-
-
-def upgrade() -> None:
-    """Upgrade schema."""
-    # ### commands auto generated by Alembic - please adjust! ###
-    pass
-    # ### end Alembic commands ###
-
-
-def downgrade() -> None:
-    """Downgrade schema."""
-    # ### commands auto generated by Alembic - please adjust! ###
-    pass
-    # ### end Alembic commands ###
