+++ app/main.py
@@ -1,7 +1,7 @@
 from fastapi import FastAPI
 from app.core.config import settings
 from app.core.middleware import setup_cors
-from app.api.router import router
+from app.api import router
 
 def create_app():
     app = FastAPI()
