+++ app/models/__init__.py
@@ -2,7 +2,7 @@ from app.models.user import User
 from app.models.poem import Poem, PoemTag, PoemLike
 from app.models.comment import Comment
 from app.models.genre import Genre
-from app.models.collection import Collection, CollectionPoem
+from app.models.collection_poem import CollectionPoem
 from app.models.tag import Tag
 from app.models.report import Report
 from app.models.notification import Notification
